consumer/requirements.txt
pymongo
transformers
torch
kafka-python